+++ csvEditorHtml/browser/CHANGELOG.md
@@ -13,6 +13,9 @@ note that the changelog is also present on the website when you click on the ver
 - added option `initialOnly_correctRowAlwaysFirstColumn` to setting `openTableAndSelectCellAtCursorPos` to only open the correct row but always column one
 	- this is now the new default
 - in the setting `openTableAndSelectCellAtCursorPos` the option `initalOnly` was renamed to `initialOnly_correctRowAndColumn`
+- added settings `lastRowOrFirstRowNavigationBehavior` and `lastColumnOrFirstColumnNavigationBehavior` to control if we wrap ad the start/end of rows/columns while navigating
+	- works with `lastRowEnterBehavior`, `lastColumnTabBehavior` (the `default` option will apply these two new settings)
+	- see https://handsontable.com/docs/6.2.2/Options.html#autoWrapRow
 
 ## 1.2.7